remove TODO comment related to checks against primary and unique keys.
I had originally thought that instead of launching a separate query for
each row, that we could beforehand do a bigger select on the current
contents of the primary and unique constrained columns of the table, and
then check the value of the current row against that dataset. The reason
was that this might improve performance. However if we did this it would
defeat the purpose of our chunking inserts to the database; i.e., the
memory usage we save by chunking would come back in the sense that we
would have to hold a good amount of the database in memory in order to
get this larger dataset to check against.

As for the small, individual queries, I don't think this should be too
much of a performance issue; but we can check.